Fixes review comments
